Revert "[Feature] Macos extension test app"
