--- SasquatchMac/SasquatchMacSwift-Extension/Base.lproj/TodayViewController.xib
@@ -1,80 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.Cocoa.XIB" version="3.0" toolsVersion="14460.31" targetRuntime="MacOSX.Cocoa" propertyAccessControl="none" useAutolayout="YES" customObjectInstantitationMethod="direct">
-    <dependencies>
-        <deployment identifier="macosx"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.CocoaPlugin" version="14460.31"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <customObject id="-2" userLabel="File's Owner" customClass="TodayViewController" customModule="SasquatchMacSwift_Extension" customModuleProvider="target">
-            <connections>
-                <outlet property="addAttachments" destination="V6r-mP-Oon" id="caI-gy-I5t"/>
-                <outlet property="extensionLabel" destination="1sU-Wy-afS" id="H9c-eI-0nu"/>
-                <outlet property="popupButton" destination="3eE-SZ-0tS" id="985-Dk-fgZ"/>
-                <outlet property="view" destination="Hz6-mo-xeY" id="0bl-1N-x8E"/>
-            </connections>
-        </customObject>
-        <customObject id="-1" userLabel="First Responder" customClass="FirstResponder"/>
-        <customObject id="-3" userLabel="Application" customClass="NSObject"/>
-        <customView simulatedAppContext="notificationCenter" translatesAutoresizingMaskIntoConstraints="NO" id="Hz6-mo-xeY">
-            <rect key="frame" x="0.0" y="0.0" width="507" height="112"/>
-            <subviews>
-                <textField horizontalHuggingPriority="251" verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="1sU-Wy-afS">
-                    <rect key="frame" x="6" y="88" width="495" height="16"/>
-                    <textFieldCell key="cell" lineBreakMode="clipping" alignment="center" title="Run #000000" id="bTn-1C-B7R">
-                        <font key="font" metaFont="cellTitle"/>
-                        <color key="textColor" name="labelColor" catalog="System" colorSpace="catalog"/>
-                        <color key="backgroundColor" name="textBackgroundColor" catalog="System" colorSpace="catalog"/>
-                    </textFieldCell>
-                </textField>
-                <button verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="V6r-mP-Oon" userLabel="Checkbox">
-                    <rect key="frame" x="6" y="64" width="495" height="18"/>
-                    <buttonCell key="cell" type="check" title="Add attachments to crashes" bezelStyle="regularSquare" imagePosition="left" state="on" inset="2" id="5Af-Y3-d6o">
-                        <behavior key="behavior" changeContents="YES" doesNotDimImage="YES" lightByContents="YES"/>
-                        <font key="font" metaFont="system"/>
-                    </buttonCell>
-                </button>
-                <button verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="Hzq-IV-DAH" userLabel="CrashMe">
-                    <rect key="frame" x="2" y="30" width="503" height="32"/>
-                    <buttonCell key="cell" type="push" title="Crash Me!" bezelStyle="rounded" alignment="center" borderStyle="border" imageScaling="proportionallyDown" inset="2" id="3hJ-SQ-2CW">
-                        <behavior key="behavior" pushIn="YES" lightByBackground="YES" lightByGray="YES"/>
-                        <font key="font" metaFont="system"/>
-                    </buttonCell>
-                    <connections>
-                        <action selector="crashMe:" target="-2" id="6hS-fk-vgb"/>
-                    </connections>
-                </button>
-                <popUpButton verticalHuggingPriority="750" translatesAutoresizingMaskIntoConstraints="NO" id="3eE-SZ-0tS">
-                    <rect key="frame" x="6" y="5" width="496" height="25"/>
-                    <popUpButtonCell key="cell" type="push" title="Crashes" bezelStyle="rounded" alignment="left" lineBreakMode="truncatingTail" state="on" borderStyle="borderAndBezel" imageScaling="proportionallyDown" inset="2" selectedItem="1C5-pp-Iha" id="0P7-Sz-HMm">
-                        <behavior key="behavior" lightByBackground="YES" lightByGray="YES"/>
-                        <font key="font" metaFont="menu"/>
-                        <menu key="menu" id="TSK-ex-IXo">
-                            <items>
-                                <menuItem title="Crashes" state="on" id="1C5-pp-Iha"/>
-                                <menuItem title="Item 2" id="3cj-OI-Z2L"/>
-                                <menuItem title="Item 3" id="SLX-i2-ql7"/>
-                            </items>
-                        </menu>
-                    </popUpButtonCell>
-                </popUpButton>
-            </subviews>
-            <constraints>
-                <constraint firstItem="1sU-Wy-afS" firstAttribute="leading" secondItem="Hz6-mo-xeY" secondAttribute="leading" constant="8" id="0za-kV-ANi"/>
-                <constraint firstItem="3eE-SZ-0tS" firstAttribute="top" secondItem="Hzq-IV-DAH" secondAttribute="bottom" constant="8" id="55j-bi-L9W"/>
-                <constraint firstItem="Hzq-IV-DAH" firstAttribute="leading" secondItem="1sU-Wy-afS" secondAttribute="leading" id="AvQ-s8-PYf"/>
-                <constraint firstItem="V6r-mP-Oon" firstAttribute="top" secondItem="1sU-Wy-afS" secondAttribute="bottom" constant="8" id="Elb-mj-xqn"/>
-                <constraint firstItem="1sU-Wy-afS" firstAttribute="top" secondItem="Hz6-mo-xeY" secondAttribute="top" constant="8" id="FhG-Ze-3qQ"/>
-                <constraint firstAttribute="trailing" secondItem="1sU-Wy-afS" secondAttribute="trailing" constant="8" id="Gee-EA-ysn"/>
-                <constraint firstItem="3eE-SZ-0tS" firstAttribute="trailing" secondItem="Hzq-IV-DAH" secondAttribute="trailing" id="IdX-1s-NGB"/>
-                <constraint firstAttribute="bottom" secondItem="3eE-SZ-0tS" secondAttribute="bottom" constant="8" id="R7f-UB-ZPc"/>
-                <constraint firstItem="V6r-mP-Oon" firstAttribute="bottom" secondItem="Hzq-IV-DAH" secondAttribute="top" constant="-8" id="XyC-6L-4WO"/>
-                <constraint firstItem="3eE-SZ-0tS" firstAttribute="leading" secondItem="Hzq-IV-DAH" secondAttribute="leading" id="YMa-H2-oLU"/>
-                <constraint firstItem="Hzq-IV-DAH" firstAttribute="trailing" secondItem="1sU-Wy-afS" secondAttribute="trailing" id="oHE-TF-z9T"/>
-                <constraint firstItem="V6r-mP-Oon" firstAttribute="leading" secondItem="Hz6-mo-xeY" secondAttribute="leading" constant="8" id="oLo-dz-hhp"/>
-                <constraint firstAttribute="trailing" secondItem="V6r-mP-Oon" secondAttribute="trailing" constant="8" id="rWo-kw-fY0"/>
-            </constraints>
-            <point key="canvasLocation" x="138.5" y="174"/>
-        </customView>
-    </objects>
-</document>
